requirements.txt
dash~=2.18.2
numpy~=2.2.3
plotly~=6.0.0
qrcode~=8.1
pandas~=2.2.3
openpyxl~=3.1.5
toolz~=1.0.0
gunicorn~=23.0.0
dash_bootstrap_components~=1.7.1
dash_extensions~=1.0
pillow~=11.0.0
mako